Reverted refactor of project related components from Layout folder. #404.
