+++ explorer/app/components/Project/components/Citation/citation.tsx
@@ -5,8 +5,8 @@ import React, { ReactNode } from "react";
 // App dependencies
 import { useConfig } from "app/hooks/useConfig";
 import { ProjectPath } from "../../common/entities";
-import { Stack } from "../../../../../common/Stack/Stack";
-import { ANCHOR_TARGET, Link } from "../../../../../Links/components/Link/link";
+import { Stack } from "../../../common/Stack/Stack";
+import { ANCHOR_TARGET, Link } from "../../../Links/components/Link/link";
 import { SectionDetailsEmpty } from "../Section/components/SectionDetailsEmpty/sectionDetailsEmpty";
 import { Section } from "../Section/section";
