+++ explorer/app/components/Project/components/Details/details.stories.tsx
