+++ explorer/app/components/Project/components/Details/details.tsx
