+++ explorer/app/components/Project/components/Hero/hero.tsx
@@ -6,16 +6,13 @@
 import React from "react";
 
 // App dependencies
+import { Stack } from "app/components/common/Stack/Stack";
 import {
   Breadcrumb,
   Breadcrumbs,
-} from "../../../../../common/Breadcrumbs/breadcrumbs";
-import {
-  Status,
-  StatusBadge,
-} from "../../../../../common/StatusBadge/statusBadge";
-import { HeroTitle, Title } from "../../../../../common/Title/title";
-import { Stack } from "app/components/common/Stack/Stack";
+} from "../../../common/Breadcrumbs/breadcrumbs";
+import { Status, StatusBadge } from "../../../common/StatusBadge/statusBadge";
+import { HeroTitle, Title } from "../../../common/Title/title";
 
 interface Props {
   breadcrumbs?: Breadcrumb[];
