+++ explorer/app/components/Project/components/Publications/publications.tsx
@@ -4,7 +4,7 @@ import React from "react";
 
 // App dependencies
 import { Publication } from "../../common/entities";
-import { ANCHOR_TARGET, Link } from "../../../../../Links/components/Link/link";
+import { ANCHOR_TARGET, Link } from "../../../Links/components/Link/link";
 import { SectionDetailsEmpty } from "../Section/components/SectionDetailsEmpty/sectionDetailsEmpty";
 import { Section } from "../Section/section";
