+++ explorer/app/components/Project/components/Section/section.styles.ts
@@ -1,7 +1,7 @@
 // TODO section styles refactor to be resolved with https://github.com/clevercanary/data-browser/issues/400
 import { SectionTitle as Title } from "./components/SectionTitle/sectionTitle";
 import styled from "@emotion/styled";
-import { BREAKPOINT } from "../../../../../../hooks/useBreakpointHelper";
+import { BREAKPOINT } from "../../../../hooks/useBreakpointHelper";
 import { ButtonBase, Typography } from "@mui/material";
 
 export const Section = styled.div`
