+++ explorer/app/components/Project/components/Section/section.tsx
@@ -11,7 +11,7 @@ import {
   BREAKPOINT,
   BREAKPOINT_FN_NAME,
   useBreakpointHelper,
-} from "../../../../../../hooks/useBreakpointHelper";
+} from "../../../../hooks/useBreakpointHelper";
 
 // Styles
 import {
