+++ explorer/app/components/Project/components/SupplementaryLinks/supplementaryLinks.tsx
@@ -4,8 +4,8 @@ import React, { Fragment } from "react";
 
 // App dependencies
 import { SupplementaryLink } from "../../common/entities";
-import { Stack } from "../../../../../common/Stack/Stack";
-import { ANCHOR_TARGET, Link } from "../../../../../Links/components/Link/link";
+import { Stack } from "../../../common/Stack/Stack";
+import { ANCHOR_TARGET, Link } from "../../../Links/components/Link/link";
 import { SectionDetailsEmpty } from "../Section/components/SectionDetailsEmpty/sectionDetailsEmpty";
 import { Section } from "../Section/section";
