+++ explorer/app/components/Project/components/TitledText/titledText.styles.ts
