+++ explorer/app/components/Project/components/TitledText/titledText.tsx
