+++ explorer/app/components/index.tsx
@@ -7,19 +7,19 @@ export { StaticImage } from "./common/StaticImage/staticImage";
 export { Cell } from "./Index/components/Cell/cell";
 export { Summaries } from "./Index/components/Hero/components/Summaries/summaries";
 export { NTagCell } from "./Index/components/NTagCell/nTagCell";
-export { Citation } from "./Layout/components/BackPage/components/Citation/citation";
-export { CollaboratingOrganizations } from "./Layout/components/BackPage/components/CollaboratingOrganizations/collaboratingOrganizations";
-export { Contacts } from "./Layout/components/BackPage/components/Contacts/contacts";
-export { Contributors } from "./Layout/components/BackPage/components/Contributors/contributors";
-export { DataCurators } from "./Layout/components/BackPage/components/DataCurators/dataCurators";
-export { DataReleasePolicy } from "./Layout/components/BackPage/components/DataReleasePolicy/dataReleasePolicy";
-export { Description } from "./Layout/components/BackPage/components/Description/description";
-export { Details } from "./Layout/components/BackPage/components/Details/details";
-export { Hero as ProjectHero } from "./Layout/components/BackPage/components/Hero/hero";
-export { Publications } from "./Layout/components/BackPage/components/Publications/publications";
-export { Section } from "./Layout/components/BackPage/components/Section/section";
-export { SupplementaryLinks } from "./Layout/components/BackPage/components/SupplementaryLinks/supplementaryLinks";
-export { TitledText } from "./Layout/components/BackPage/components/TitledText/titledText";
+export { Citation } from "./Project/components/Citation/citation";
+export { CollaboratingOrganizations } from "./Project/components/CollaboratingOrganizations/collaboratingOrganizations";
+export { Contacts } from "./Project/components/Contacts/contacts";
+export { Contributors } from "./Project/components/Contributors/contributors";
+export { DataCurators } from "./Project/components/DataCurators/dataCurators";
+export { DataReleasePolicy } from "./Project/components/DataReleasePolicy/dataReleasePolicy";
+export { Description } from "./Project/components/Description/description";
+export { Details } from "./Project/components/Details/details";
+export { Hero as ProjectHero } from "./Project/components/Hero/hero";
+export { Publications } from "./Project/components/Publications/publications";
+export { Section } from "./Project/components/Section/section";
+export { SupplementaryLinks } from "./Project/components/SupplementaryLinks/supplementaryLinks";
+export { TitledText } from "./Project/components/TitledText/titledText";
 export { Text } from "./Text/Text";
 export { TextLinks } from "./TextLinks/TextLinks";
 export { ValueBlock } from "./ValueBlock/ValueBlock";
